refactor(ipc): IPC 共通ライブラリをワークスペースパッケージに分離
- IPC 共通ライブラリを packages/ipc へ移設し、desktop 側の参照を @repo/ipc に切り替え
- packages/ipc に package.json、tsconfig、oxlint 設定を追加して、独立したワークスペースパッケージとして定義
- ルートの test と lint スクリプトを再帰実行に変更し、新しい IPC パッケージも対象に含める
