+++ packages/ipc/.oxlintrc.json
@@ -7,7 +7,7 @@
   "env": {
     "builtin": true
   },
-  "ignorePatterns": ["**/*.d.ts", "**/dist-electron/**", "**/coverage/**", "**/dist/**"],
+  "ignorePatterns": ["**/*.d.ts", "**/coverage/**"],
   "rules": {
     "import/no-duplicates": "error",
     "typescript/consistent-type-imports": [
@@ -18,18 +18,5 @@
       }
     ],
     "import/consistent-type-specifier-style": ["error", "prefer-top-level"]
-  },
-  "overrides": [
-    {
-      "files": ["src/**/*.{ts,tsx}"],
-      "plugins": ["react", "typescript", "unicorn", "oxc", "import"],
-      "jsPlugins": ["eslint-plugin-use-no-memo"],
-      "rules": {
-        "use-no-memo/react-hook-form": "error",
-        "use-no-memo/tanstack-table": "error",
-        "react/rules-of-hooks": "error",
-        "react/exhaustive-deps": "warn"
-      }
-    }
-  ]
+  }
 }
