+++ packages/ipc/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@repo/ipc",
+  "version": "0.0.1",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "import": "./src/index.ts"
+    },
+    "./main": {
+      "types": "./src/main/index.ts",
+      "import": "./src/main/index.ts"
+    },
+    "./browser": {
+      "types": "./src/browser/index.ts",
+      "import": "./src/browser/index.ts"
+    }
+  },
+  "scripts": {
+    "test": "tsgo --noEmit && vitest run",
+    "lint": "tsgo --noEmit && oxlint"
+  },
+  "devDependencies": {
+    "@typescript/native-preview": "7.0.0-dev.20260418.1",
+    "electron": "^41.2.1",
+    "oxlint": "^1.60.0",
+    "typescript": "^6.0.3",
+    "vitest": "^4.1.4"
+  },
+  "packageManager": "pnpm@10.33.0"
+}
